REF: Factor out string methods from ArrowExtensionArray
